refactor wpFragments(static getInstance, reorganize launcher recycler classes, fab -> fragments, new app icon)
