+++ app/src/main/res/layout/fragment_recycler_launcher.xml
@@ -1,22 +1,35 @@
-<android.support.v7.widget.RecyclerView
-    android:id="@+id/launcher_recycler_view"
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-</android.support.v7.widget.RecyclerView>
-<!--<?xml version="1.0" encoding="utf-8"?>-->
-<!--<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"-->
-    <!--android:layout_width="match_parent" android:layout_height="match_parent"-->
+    android:orientation="vertical"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
     <!--android:id="@+id/coordinator_parent_recycler">-->
-    <!--<android.support.v7.widget.RecyclerView-->
-        <!--android:id="@+id/launcher_recycler_view"-->
-        <!--android:layout_width="match_parent"-->
-        <!--android:layout_height="match_parent">-->
-    <!--</android.support.v7.widget.RecyclerView>-->
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/launcher_recycler_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+    </android.support.v7.widget.RecyclerView>
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/fab_launcher"
+        android:layout_margin="@dimen/margin"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_anchor="@id/launcher_recycler_view"
+        app:layout_anchorGravity="bottom|end"
+        android:src="@drawable/ic_add_white_48dp"/>
+</android.support.design.widget.CoordinatorLayout>
+    <!--<android.support.v7.widget.RecyclerView android:id="@+id/launcher_recycler_view"-->
+    <!--android:layout_width="match_parent"-->
+    <!--android:layout_height="match_parent"-->
+    <!--xmlns:android="http://schemas.android.com/apk/res/android"-->
+    <!--xmlns:app="http://schemas.android.com/apk/res-auto">-->
     <!--<android.support.design.widget.FloatingActionButton-->
         <!--android:id="@+id/fab_launcher"-->
         <!--android:layout_margin="@dimen/margin"-->
         <!--android:layout_width="wrap_content"-->
         <!--android:layout_height="wrap_content"-->
-        <!--android:layout_gravity="bottom|end"/>-->
-<!--</android.support.design.widget.CoordinatorLayout>-->
+        <!--app:layout_anchor="@id/launcher_recycler_view"-->
+        <!--app:layout_anchorGravity="bottom|end"-->
+        <!--android:src="@drawable/ic_add_white_48dp"/>-->
+<!--</android.support.v7.widget.RecyclerView>-->
